Session: Add session a store depending upon the config
