+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "connect-mongo": "^3.2.0",
     "convict": "^6.0.0",
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
