+++ server/conf/config/schemaMongodb.json
@@ -32,5 +32,19 @@
                 "env": "DB_PORT1"
             }
         }
-    ]
+    ],
+    "sessions": {
+        "store": {
+            "doc": "Store for sessions",
+            "format": "String",
+            "default": "mongodb",
+            "env": "STORE_NAME"
+        },
+        "maxAge" : {
+            "doc": "Maxage for session",
+            "format": "Number",
+            "default": 900000,
+            "env": "STORE_NAME"
+        }
+    }
 }
