+++ server/conf/util.js
@@ -1,0 +1,27 @@
+const { v4: uuidv4 } = require('uuid');
+
+var setSessionConfig = function (conf, session, dbs) {
+
+    let sessionConfig = {
+        genid: (req) => { //The function is given req as the first argument
+            console.log("insid middleware \n session id " + req.sessionID);
+            return uuidv4();
+        },
+        secret: "test",
+        resave: false,
+        saveUninitialized: true,
+        maxAge: conf.get("sessions.maxAge")
+    };
+
+    if (conf.get("sessions.store") == "filestore") {
+        const FileStore = require('session-file-store')(session);
+        var fileStoreOptions = {};
+        sessionConfig.store = new FileStore(fileStoreOptions)
+    } else {
+        const MongoStore = require('connect-mongo')(session);
+        sessionConfig.store = new MongoStore({ mongooseConnection: dbs.login });
+    }
+    return sessionConfig;
+};
+
+module.exports.setSessionConfig = setSessionConfig;
